[feat] 모집글에 지원한 사람이 있으면 모집글 삭제 시도시 에러코드 전송 (#6)

+++ Gathering_be/src/main/java/com/Gathering_be/service/ProjectService.java
@@ -85,6 +85,10 @@ public class ProjectService {
         Project project = findProjectById(projectId);
         validateMemberAccess(project);
 
+        if (applicationRepository.existsByProjectId(projectId)){
+            throw new ProjectHasApplicantsException();
+        }
+
         project.getProfile().removeProject(project.isClosed());
         projectRepository.deleteById(projectId);
     }
